# Inline `ASTNode` bindings dependencies and observers