--- a/src/compiler/crystal/semantic/bindings.cr
+++ b/src/compiler/crystal/semantic/bindings.cr
@@ -1,7 +1,101 @@
 module Crystal
+  # Specialized container for ASTNodes to use for bindings tracking.
+  #
+  # The average number of elements in both dependencies and observers is below 2
+  # for ASTNodes. This struct inlines the first two elements saving up 4
+  # allocations per node (two arrays, with a header and buffer for each) but we
+  # need to pay a slight extra cost in memory upfront: a total of 6 pointers (48
+  # bytes) vs 2 pointers (16 bytes). The other downside is that since this is a
+  # struct, we need to be careful with mutation.
+  struct SmallNodeList
+    include Indexable(ASTNode)
+
+    @first : ASTNode?
+    @second : ASTNode?
+    @tail : Array(ASTNode)?
+
+    def size
+      size = 0
+      size += 1 unless @first.nil?
+      size += 1 unless @second.nil?
+      if tail = @tail
+        size += tail.size
+      end
+      size
+    end
+
+    def unsafe_fetch(index : Int)
+      if first = @first
+        if index == 0
+          return first
+        else
+          index -= 1
+        end
+      end
+      if second = @second
+        if index == 0
+          return second
+        else
+          index -= 1
+        end
+      end
+      if tail = @tail
+        tail.unsafe_fetch(index)
+      else
+        raise IndexError.new
+      end
+    end
+
+    def each(& : ASTNode ->)
+      if first = @first
+        yield first
+      end
+      if second = @second
+        yield second
+      end
+      if tail = @tail
+        tail.each { |node| yield node }
+      end
+    end
+
+    def push(node : ASTNode) : self
+      if @first.nil?
+        @first = node
+      elsif @second.nil?
+        @second = node
+      else
+        tail = @tail ||= Array(ASTNode).new
+        tail.push(node)
+      end
+      self
+    end
+
+    def reject!(& : ASTNode ->) : self
+      if first = @first
+        if yield first
+          @first = nil
+        end
+      end
+      if second = @second
+        if yield second
+          @second = nil
+        end
+      end
+      if tail = @tail
+        tail.reject! { |node| yield node }
+      end
+      self
+    end
+
+    def concat(nodes : Enumerable(ASTNode)) : self
+      nodes.each { |node| self.push(node) }
+      self
+    end
+  end
+
   class ASTNode
-    property! dependencies : Array(ASTNode)
-    property observers : Array(ASTNode)?
+    getter dependencies : SmallNodeList = SmallNodeList.new
+    getter observers : SmallNodeList = SmallNodeList.new
     property enclosing_call : Call?
 
     @dirty = false
@@ -107,8 +201,8 @@ module Crystal
     end
 
     def bind_to(node : ASTNode) : Nil
-      bind(node) do |dependencies|
-        dependencies.push node
+      bind(node) do
+        @dependencies.push node
         node.add_observer self
       end
     end
@@ -116,8 +210,8 @@ module Crystal
     def bind_to(nodes : Indexable) : Nil
       return if nodes.empty?
 
-      bind do |dependencies|
-        dependencies.concat nodes
+      bind do
+        @dependencies.concat nodes
         nodes.each &.add_observer self
       end
     end
@@ -130,9 +224,7 @@ module Crystal
         raise_frozen_type freeze_type, from_type, from
       end
 
-      dependencies = @dependencies ||= [] of ASTNode
-
-      yield dependencies
+      yield
 
       new_type = type_from_dependencies
       new_type = map_type(new_type) if new_type
@@ -167,8 +259,12 @@ module Crystal
       nodes.each &.remove_observer self
     end
 
+    def unbind_from(nodes : SmallNodeList)
+      @dependencies.try &.reject! { |dep| nodes.any? &.same?(dep) }
+      nodes.each &.remove_observer self
+    end
+
     def add_observer(observer)
-      observers = @observers ||= [] of ASTNode
       observers.push observer
     end
 
@@ -269,7 +365,7 @@ module Crystal
       visited = Set(ASTNode).new.compare_by_identity
       owner_trace << node if node.type?.try &.includes_type?(owner)
       visited.add node
-      while deps = node.dependencies?
+      while deps = node.dependencies
         dependencies = deps.select { |dep| dep.type? && dep.type.includes_type?(owner) && !visited.includes?(dep) }
         if dependencies.size > 0
           node = dependencies.first