@ggiraldez 2024-10-19
Damn... I wished I saw [this](https://github.com/crystal-lang/crystal/issues/12851) before I even started 😬

The CI is failing anyway, and maybe the reason is the same as the last time. Will investigate further.