@straight-shoota 2024-10-19
Ah, I assumed you were aware of that. 🙈
I had definitely mentioned there had been a previous attempt that looked nice but didn't work out.

Anyway the issue here must be something relatively obvious. I presume it shouldn't be hard to fix. The interpreter failure was more subtle.