@ggiraldez 2024-10-19
> Ah, I assumed you were aware of that. 🙈 I had definitely mentioned there had been a previous attempt that looked nice but didn't work out.

No worries. It was pretty straightforward to implement. I know that you mentioned it, but it didn't occur to me that there was a trail to follow.
 
> Anyway the issue here must be something relatively obvious. I presume it shouldn't be hard to fix. The interpreter failure was more subtle.

Not sure it's obvious 😬 So far, I spent more time trying to figure out what's broken than actually implementing the optimisation. The good news is that this doesn't reproduce #12769.