@straight-shoota 2024-10-19
Turning `SmallNodeList` into a class seems to fix the failure. So it's a pass-by-value issue.
We can dig through the source code to finde where we pass an instance. Or we could keep it a class, and maybe inline the allocation with `ReferenceStorage` 🤷